+++ agent/agent-bootstrap/src/main/java/com/microsoft/applicationinsights/agent/bootstrap/configuration/Configuration.java
@@ -29,4 +29,7 @@ public class Configuration {
 
     // transient so that Moshi will ignore when binding from json
     public transient Path configPath;
+
+    // transient so that Moshi will ignore when binding from json
+    public transient Long lastModifiedTime;
 }
